New interface for minimization; changes in modules organization
Minimization now have a proper argument parser; old utils.py is now moved into physics.py, which is more consistent with content. Need to change "globals" to ALL CAPS; some functions in run_minimize.py are now moved into the new utils.py - to be done the same for train.py and run_evo.py; SAVE ARGS FUNCTION TO BE IMPLEMENTED; TESTS NOT PERFORMED YET
